fixing some -Wsign-compare stuff

+++ libraries/random/random.c
@@ -65,9 +65,9 @@ rnd_pcg32_rand_range_fill(uint32_t *mem, uint32_t lim, size_t n) {
 }
 
 void
-rnd_pcg32_rand_uniform_fill_float(float *mem, uint32_t n) {
+rnd_pcg32_rand_uniform_fill_float(float *mem, size_t n) {
     ensure_initialized();
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         uint32_t r = rand_int();
         mem[i] = (float)r / (float)4294967296;
     }
